none working version of part 2 of 9th of december

+++ 9December/solution.h
@@ -2,5 +2,5 @@
 #define SOLUTION9_H
 #include <bits/stdc++.h>
 uint64_t solution9();
-int second_solution9();
+uint64_t second_solution9();
 #endif
